Initial cycle breaking implementation
Summary:
This diff is the first one attempting to actually improve cycle breaking
determinism.

The algorithm we use should iliminate direct nondeterminism coming from
scenarios where threads can race to determine where the recursive result
is actually produced in a cycle: we now always produce it at the minimal
idx of the cycle.

There are still data races for now though: we are not yet forcing the
recursive `Var`s at any predetermined place, so if they remain unforced
they are exposed to arbitrary data races. And even if we forced them,
as long as we are writing answers that contain unforced `Var`s in to
the final `Calculation` result, we are allowing threads to race on
pinning those vars. I will address these problems later, for now the goal
is just to break cycles in a fixed place.

**Summary of test changes**

The main change is that all of the tests whose results inexplicably changed
in D75722156 revert to old behavior; for the most part the old behavior is better;
in particular loops are considerably better behaved after deterministically breaking
on the lowest-idx key (which will generally be the phi, this is the right place to break).

One of the things that alerted me to cycle breaking being a problem was that
a trivial loop broke when I introduced `Pin` bindings, which was only possible
if the ordering was critical. That test is now passing again, presumably because
we now solve the `Phi` first which is as it should be :)

The `var_expansion` lsp test is essentially the same one (I copied it to a unit
test when it broke, specifically because I thought we might need to debug
tricky ordering issues). It is likewise now fixed :)

All of the `super` tests that changed in D75722156 (which I couldn't explain,
but were another indication to me that ordering is a bigger problem than I
had realized prior to that diff) have now reverted to the previous behavior.
This means we trade some bugs for other bugs, but I think it's probably a
net win to get back where we started with these.

**No Conformance Changes**

This is good - except on highly cyclic code this is supposed to be either a
no-op or (in the case of loops in particular) an improvement.

Reviewed By: ndmitchell

Differential Revision: D76948127

fbshipit-source-id: 93afa3a7a601447962a7fbaba78c80be72dcd5a4