Rename cycle detected / completed handlers
Summary:
I realized the names of these funcitons sort of sound like they
might be boolean predicates. This is particularly problematic in
the case of `cycle_detected`, which actually *does* return a boolean,
but the boolean does not tell us "was a cycle detected?" but rather
"should we immediately break the cycle that was just detected?".

I think clarifying the names with a `handle_` prefix is worth the
verbosity.

Reviewed By: yangdanny97, ndmitchell

Differential Revision: D77490600

fbshipit-source-id: 6cee3642f3719dcfefd73b16693cfc7cc1c7d7f3