Un-mark a pytorch TODO
Summary:
It appears to me this is working as expected. I tweaked the example so it's more
obvious (using Optional[Any] makes it slightly harder to see whether narrowing is actually
happening than using a concrete type). I also renamed it, since it isn't really an
attribute-related narrowing test at all - there's an attribute involved, but the
narrow is on as subscript.

Reviewed By: migeed-z

Differential Revision: D77551063

fbshipit-source-id: a5f22ce9fe1ff7b90e654d98612e0bac267fe130