Create a struct for thread-local AnswersSolver info
Summary:
THis is a nice BE improvement already, since we were always passing the calculation
stack and the cycle stack together, but it will also give us a place to put
thread-isolated storage.

This is important because we need to avoid storing `Var`s in `Calculation`, which is
fully public, in order to eliminate data races pinning the `Var`s.

Reviewed By: ndmitchell

Differential Revision: D77565437

fbshipit-source-id: b025030b98cf4000684c4a82ccf6bd36d23c6c37