Add a mechanism to know if we've emptied the cycle stack
Summary:
We want to work on isolating `Var`s in the current thread to eliminate data races pinning
them. This will require a way to store some or all answers, when we are handling cycles,
in thread-local storage.

But we still need a way to know when it is time to finalize the answers (at which point
we will force any Vars that are still unpinned) and write them out.

We can't do that when we finish a cycle, because in a densely-connected binding graph there
can be many cycles, this is why `Cycles` is built around a cycle *stack* rather than a single
cycle.

Here we make `on_cycle_completed` indicate in the return whether the cycle stack is now
empty (which means we just finished solving a strongly-connected component).

Reviewed By: ndmitchell

Differential Revision: D77565436

fbshipit-source-id: 7e153cc00dda9503a528c2c25e2f0981796a03c2