Update `continue_find_module` func signature to return `Result` to prep for error handling
Summary:
Updates the `continue_find_module` function signature to return a Result type, allowing for more robust error handling.

Fixes calling code to use unwrap calls to propagate errors, which will be replaced with proper error handling in a future diff when the error implementation happens. This change is a precursor to improving how we handle compiled modules and ignoring errors when no stubs are available.

Reviewed By: grievejia, connernilsen

Differential Revision: D77555853

fbshipit-source-id: 4862210e81024862795fc3be2b3299b1d1c79f6b