Implement `Debug` for `CalcId`
Summary:
It's occasionally useful to be able to debug with no `AnswersSolver`
to display (in particular, inside of `Cycle` construction).

Note: I think we may actually be able to also implement `Display`; the
AnswersSolver is just using `Bindings` these days, so that logic can
probably move from `debugging.rs`. But that's a bigger change and I'm
not blocked on it, so I'll tackle it later.

Reviewed By: ndmitchell

Differential Revision: D77610540

fbshipit-source-id: ca01b23a642b8952ece677da812f92f1821e6de3