Return environment finding errors instead of printing to console immediately
Summary:
Building off the two previous diffs, since we can now report errors from  `ConfigFile::configure()`, we should instead return them from various `PythonEnvironment` functions rather than logging them immediately. We also add an error if we can't find an interpreter when we expect to.

With this, we now don't have to worry about messing up the progress bar, and a failure to query an interpreter will help ensure we return with a non-zero exit code (this is fine, since it most likely indicates the config is misconfigured).

Reviewed By: grievejia

Differential Revision: D77618820

fbshipit-source-id: ed4dce3ca5923407f3b4617d83eab2f8e2ab57fd