Remove CalleeKind::DataclassTransformDecorator
Summary: The only reason I added this CalleeKind was to try to force DataclassTransformDecorator to fit nicely into the `match` on callee kind that we do when processing decorators. It's not strictly necessary and makes the next change I want to make more complicated than it needs to be, so remove it.

Reviewed By: ndmitchell

Differential Revision: D77636322

fbshipit-source-id: 378453ca5062f2a43242f2baac45417d59e7c104