Slightly refactor decorator loop in function.rs
Summary:
Merges all checks into the match on callee kind, moves cases that don't use the callee kind to the bottom, and uses `_` for said cases.

This is still sort of weird because of the multiple "_ if ..." cases, but it's at least obvious now what all the cases are and which ones use the callee kind.

Reviewed By: ndmitchell

Differential Revision: D77636318

fbshipit-source-id: d4d4c1c30b2bc7e69c48189bc1c7f6ad4e8b732f