Make `completions()` a wrapper around `completions_no_union_intersection()`
Summary:
whats_in_a_name

The pre-existing auto-complete API `completions()` only works with non-union and non-intersection type. Therefore I've renamed the function `completions_no_union_intersection()` to clarify that's the case.

I'm still keeping the original `completions()` as a thin wrapper around `completions_no_union_intersection()` for now to minimize downstream impact. In the future, we may want to extend `completions()` to handle union and intersection types in a certain way (I've commented on what "certain way" means in the code)

Why is it necessary to make a distinction between `completions()` vs. `completions_no_union_intersection()`? This is because while working on D77617295, I realized that our pre-existing goto-def logic for attribute accesses depends on `completions()`. This is, technically speaking, inaccurate because what we really want to depend on is the non-union-non-intersection variant of the autocompletion API, due to the two clients' difference in how unions should be resolved:
- For autocompletion, if `x` is of type `A | B`, then the attribute name `y` should be included in the result only if `y` exists as an attribute on BOTH `A` and `B`.
- For gotodef, if `x` is of type `A | B`, then gotodef on attribute name `y` feels like it should be included in the results as long as `y` exists as an attribute on EITHER `A` or `B` -- This could be a debatable position to take, but I'm going with it for now as I think it's the slightly more useful behavior.

Having 2 different versions of auto-complete API makes my intention a bit more clear.

Reviewed By: kinto0

Differential Revision: D77617297

fbshipit-source-id: 724bf3809fd4ee2f52282011e964e08ab1de2ebd