Change the return type of `find_definition()` from an option to a vec
Summary:
multiple_hands_typing

This stack attempts to add support for resolving a single gotodef request to multiple definitions. To achieve that, we'll need to adjust the return types of various APIs, and change them from returning an `Option` to a `Vec`.

This diff performs the return type adjustment for the inner layer of those APIs: `find_definition()`. Below the API, we simply convert the `Option` to a singleton list atm (so this diff could preserve pre-existing semantics). On the caller side, I only add logic to utilize the new `Vec` to two clients: `goto_definition()` (which has already been changed to support multi definition in D77617300), and find local references (i.e. now find local reference for a given identifier `x` will include references to all possible definitions of `x`). Rest of the clients are not super relevant to my immediate goal and therefore I'd like to postpone changing them.

Reviewed By: kinto0

Differential Revision: D77617299

fbshipit-source-id: fe25fd794bb58239dfbca66127754f71a0a36abf