Support goto multiple definitions for attribute accesses
Summary:
multiple_hands_typing

This stack attempts to add support for resolving a single gotodef request to multiple definitions.

One such scenario is when we gotodef on an attribute that are defined in multiple classes. e.g.

```
class Foo:
  x = 0
class Bar:
  x = 1

def test(u: Foo | Bar) -> None:
  u.x
#   ^
```
All I did here was to adjust the return type of `find_definition_for_attribute()`, and within the function, we check if the base type is a union/intersection, perform gotodef on each of the element type, and aggregate the result back.

Reviewed By: kinto0

Differential Revision: D77617295

fbshipit-source-id: d474ada33f03b105c6eb3d56d29d7cd94261251b