Change the return type of `goto_definition()` from an option to an option vec
Summary:
multiple_hands_typing

This stack attempts to add support for resolving a single gotodef request to multiple definitions. To achieve that, we'll need to adjust the return types of various APIs, and change them from returning an `Option` to a `Vec`.

This diff performs the return type adjustment for the outmost layer of those APIs: `goto_definition()`. Below the API, we simply convert the `Option` to a singleton list atm (so this diff could preserve pre-existing semantics). On the caller side, I only add logic to utilize the new `Vec` to one single client: the goto-def resolver in our IDE language server. Other clients, including type traces and playground language server, still only support 1 definition -- extending those clients is slightly out of scope of my immediate goal so I'll have to defer that for now.

Reviewed By: kinto0

Differential Revision: D77617300

fbshipit-source-id: 6a4e4106bb54f15e296ca71f223ef744a66a02b9