attempt 2 at fixing vararg param subset
Summary:
does what D77273848 is trying to do without causing an infinite loop

fixes https://github.com/facebook/pyrefly/issues/569 + gives a small conformance improvement

the control flow in this function is very complicated, since there's a loop with two iterators that we're manually stepping + breaks and early returns, and the state of the iterators when we exit the loop affects the logic afterwards. the debugger was very useful here.

Reviewed By: ndmitchell

Differential Revision: D77596204

fbshipit-source-id: 7a6a9b060d6f18bfb963b921ad96dd5699384e5b