only create a definition for AugAssign if the LHS is a Name
Summary:
We shouldn't be creating a definition for the LHS of aug assignments since the name should already exist. However, it's currenty required by the bindings step when the LHS is a name.

https://www.internalfb.com/code/fbsource/[7b8597b3ac1a995929f7078a69a0937f530dabc4]/fbcode/pyrefly/pyrefly/lib/binding/stmt.rs?lines=399

This diff changes it so that we only make the definition if the LHS is a name, instead of making extra definitions for invalid targets like Expr::Star (which is the cause of the panic in the linked bug).

fixes https://github.com/facebook/pyrefly/issues/454

Reviewed By: stroxler

Differential Revision: D76752577

fbshipit-source-id: a5bbc7cc563ae3068cc5183ea5c332cdd705d4f6