Merge BoolKeywords and DataclassKeywords
Summary:
Merges BoolKeywords and DataclassKeywords into a revamped DataclassKeywords struct that follows the pattern from the previous diffs of having dedicated fields for dataclass parameters.

This also required me to tweak how we create dataclass keywords for dataclass transforms, as we previously relied on being able to tell when a keyword was unset (absent from BoolKeywords).

Reviewed By: ndmitchell

Differential Revision: D77714442

fbshipit-source-id: e1014bf692dc423f6bb1153d6e8c396870a9ebb5