Rework cycle stack construction.
Summary:
I realized there's a problem in our cycle handling: the handling of
the case where `break_at == detected_at` isn't exactly the same as
the normal case, sometimes the break was winding up in the wrong spot.

This doesn't matter on trunk today because we never actually depend on
knowing when the cycle finishes, so everything just works even though
we don't accurately determine when cycles should pop, but once we
try to isolate vars per-thread it becomes *very* important to know when
the cycle is finished!

As of this diff, we always make sure that `break_at` is at the bottom of the
unwind stack, and is not part of the recursion stack at all.

This logic is tricky to get right, so this diff adds validation that makes
it much clearer things are working as intended (and also makes it much
more explicit what actually *is* intended!) by asserting that the bottom
of the `unwind_stack` is always `break_at`.

Reviewed By: ndmitchell

Differential Revision: D77621159

fbshipit-source-id: c98c67a4624da913e43afe67917c296eeb87a11d