Move inner `post_calculate_state` to `Cycle`
Summary:
Same rationale as the earlier diff: the Cycle logic isn't quite right at the moment,
recursion always works (which is the only thing our existing Calculation-based
setup depends on) but unwinding does not for sufficiently complex graphs.

It's easier to see what's happening if the constructor and mutators for `Cycle` all
live in one spot, and the `Cycles` layer is only responsible for stack-level semantics.

This is the function where we actually have some problems: it's not really correct to
run the post-calculation handler only for the top of the stack, that assumes cycles
behave like simple examples that are easy to think of. The cycles are emergent from
a depth-first search of the calculation graph, which means it's entirely possible to
encounter new cycles as we unwind that might be either subsets of an already-active
cycle or reach further back in the `CalcStack`; in either case, the current logic
is wrong, we need to be checking *all* active cycles. This is part of my motivation
for extracting the code, since we need to re-wire things and it is a lot clearer
at that point for the stack-level and cycle-level logic to be cleanly layered.

Reviewed By: ndmitchell

Differential Revision: D77664940

fbshipit-source-id: cb892d77a7ad4f58bea67b4a3b60b396b8ee29ba