Always update the unwind stack for cycle participants
Summary:
The pre-calculate cycle state check is responsible for moving the current calculation
from the recursion stack to the unwind stack if we are at a cycle participant.

This means that if we short-circuit the cycle because another thread already provided
us with a value, we still need to update the unwind stack.

This logic is a bit messy - it might be cleaner to flip the order of the calculation
versus cycle state checks, but I'd have to think carefully about that and I first
want to get to a point where the existing setup is correct.

Reviewed By: ndmitchell

Differential Revision: D77669455

fbshipit-source-id: a7a0e0ab83ec5ca3d78ade2ab124d0d8c6062a86