Inject globals earlier
Summary: Previously for things like `__name__` we injected them on use. That means that if no one uses `__name__` it didn't exist, and if anyone did `module.__name__` it wasn't there. That isn't how things really work, so instead inject the globals earlier on in the process, unconditionally. Now lookups happen normally, and they get exported so you can see them in the result.

Reviewed By: migeed-z

Differential Revision: D77730584

fbshipit-source-id: 718c4a8c86c75f3429e23ac6da4bd7a5da39c831