unwind all (applicable) cycles on calculation completed
Summary:
It is entirely possible for cycles to overlap. This cannot occur during the recursion
phase because we know based on how detection of a cycle works that it is the first cycle
we'll encounter via depth-first search. But during the unwind step we traverse the full
breadth of the graph, and we can encounter sub-cycles or cycles that go further "back"
in the `CalcStack`; in either case, we may be unwinding multiple cycles at once when
we finish a calculation that participates in multiple cycles.

----------------------

Just to illustrate how these kinds of overlapping cycles may occur, imagine we
encounter a cycle when the current calculation stack is
```
2 8 4 6 9
```
and we encounter a dependency from 9 to 8. We'll recurse outward until the stack
becomes
```
2 8 4 6 9 8 4
```
and then break and start unwinding.

But what happens if we unwind back to 9 and encounter a dependency on 6? Now we
have a sub-cycle, we have to create another recursive placeholder for 6. When we
finish computing 9 using that recursive placeholder, we'll have two cycles on the
stack that both include 9; we need to update both of them not just the top one.

A similar issue can occur with a new cycle that goes further back in the stack;
for example if we detect a dependency from 6 to 2 we'll get a new cycle that
in some sense supersedes the one we are currently unwinding.

What this really amounts to is that the cycle "stack" is really only clearly
behaving like a "stack" during the recursion phase; in the unwinding phase it
is more like just a heap of cycles in the most general case.

-----------------------

Note: Cycle::on_calculation_finished is much more complex than it needs to be
now, but I wanted to keep this diff as focused as possible and the existing logic
still works. I clean things up in the next diff.

Reviewed By: ndmitchell

Differential Revision: D77669454

fbshipit-source-id: ff3c5daca22747c875494eb13b535d57ddd482c6