Add an `unwound` vec to `Cycle` for debugging
Summary:
Once we've popped calculations from a `Cycle`, we lose the ids
entirely. This is fine for production, but it makes debugging tricky
because we can't reliably see what was in the cycle.

If we instead move the items from the unwind stack to another `Vec`,
it is much easier to debug. This does come with some runtime cost, but
it's a `Vec` per cycle which shouldn't be very expensive, and we
can delete it whenever we are satisfied with our cycle handling, I think
it might be worth having for now.

Reviewed By: ndmitchell

Differential Revision: D77761339

fbshipit-source-id: 2220fab3a8be40526b04fcea365b3e0a19e747cd