Avoid cycles when inferring return types
Summary:
We've generally seen performance is much better if you don't infer return types, which is super weird, because we infer types for everything anyway, and return types aren't special. I think the root cause is the line changed in this diff. Given a function like:

```
def f(x):
    return x
```

We end up having the parameter `x` depend on the type of the function, which depends on the return type, which depends on the function (cycle). As you add parameters you very quickly build larger and larger overlapping cycles, which can cause the stack usage and cycles to increase massively. I've found examples where this change reduces the stack uage by 20x.

This change still allows us to pass context down from `x` in to the function (e.g. lambda or class functions) but doesn't allow the function to influence the type of its parameters. Which is much more correct.

Adjusts the approach in D70050192.

Reviewed By: grievejia

Differential Revision: D77800883

fbshipit-source-id: ff863adac42fd896d9999fc21aa10a4bcaea634f