Don't clone `x.subject`
Summary:
The previous diff simplified match subject extraction by getting logic that didn't
belong in the case loop out. I mentioned that this was probably more efficient, but
it cost an extra clone - with this diff we're now strictly more efficient: there
was never a need to clone `x.subject` ahead of time, we can pass it by reference.

And there's a significant readability bonus, we have a bunch of local variables
named somethign like `subject` which I was finding made this hard to read; inlining
one of them helps :)

Reviewed By: ndmitchell

Differential Revision: D77835700

fbshipit-source-id: ee5d5723eb80e68c8174ae288355698899e7bff1