A couple more tweaks to `MatchMapping`
Summary:
The name `subject_for_key` isn't exactly bad, it's the "new subject" representing
what we have after applying the key to the "outer subject".

But I kept getting confused reading it, mostly I think because I tend to read
`subject_for_key` as "the subject for some Idx<Key>" which is not what's intended.

At least to me, `subject_at_key` reads more clearly as "I'm taking the subject and
applying a key facet to it, which gives me a new subject for a recursive call to
`bind_pattern`.

Also, I moved up the `insert_binding_current` call - I think it's good practice to
avoid interleaving unrelated logic in between defining a `CurrentIdx` and binding
it; being disciplined about this minimizes the chance that we wind up passing the
wrong "current" as our `usage` at some point.

Reviewed By: ndmitchell

Differential Revision: D77835846

fbshipit-source-id: 8c98599667fd6c163eae750b98bbb0942fb2bd9e