Account for walrus in `definitions.rs` (part 2)
Summary:
I don't like using an `_` case for something as relatively small and as critical
as a statement match, it's too easy to miss cases.

For walrus the only arguably important case we were missing was `assert`, and
`raise` which have expressions in them and might continue to execute (in the case
of `raise`, the `except` and `finally` clauses could in principle be relying
on names defined inside the raise statement).

In theory a walrus can appear in a `return` or `delete`, but I think in
all these cases it's probably okay to omit becasue we're going to exit normal
control flow immediately anyway, and we do still have a hook where we force the
name into the static scope in `bindings` anyway because we need that for
comprehensions, so we're not going to panic or anything like that if we hit
these cases.

Enumerating the statements also highlights questionable-looking `TypeAlias`
logic. The branch I had to add - a type alias with no name - doesn't actually
matter because it's a syntax error, but we probably should handle this
differently. I'll clean it up in the next diff.

Reviewed By: ndmitchell

Differential Revision: D77845457

fbshipit-source-id: 245e922518b401fc882bfe4368ce4cb8d16f9381