Improve `definitions` handling of `TypeAlias`
Summary:
This was another missing case from the unit test, and given that it's not 100%
clear how the match behaves just from skimming the code, this is pretty important.

In addition, the use of a match guard made control flow look potentially buggy,
using an `if` clause inside an unguarded match feels clearer.

Finally, in principle there could be a walrus in the type alias value; added logic
and another line in the walrus unit test covering this.

Reviewed By: ndmitchell

Differential Revision: D77845456

fbshipit-source-id: 98f63be2b2dee4f0d82774e093d2985e8e6975b6