Make isinstance(_, type[Any]) work
Summary: You can pass a value of `type[Any]` as the second argument to `isinstance`, so allow that. What Pyright additionally does is ban the literal Any. It tracks that in the type system. We don't - both just have the type `type[Any]`. I'll add that as a special case in the next diff.

Reviewed By: grievejia

Differential Revision: D77887756

fbshipit-source-id: 9e1f62e53a13ea4a968712454b7ec24a14c8caa1