Force recursive vars when a cycle finishes
Summary:
Today, a cycle can finish without pinning the recursive var used to
break it. This means the unpinned Var can be hanging around in the results
indefinitely, exposing us to a very long time window for different exports
that are not part of the cycle but downstream of it to potentially pin the Var
to different values.

This diff changes Pyrefly to always force the var (which will produce
either a default type or a Any if the type has not yet been pinned) as soon
as we've gone all the way around the cycle and gotten back to the 'break at'.

We still do not isolate the results per thread, so there are still data
races possible while a cycle is executing (another thread can read a cached
result that uses a Var produced along the cycle before the cylcle has
completed). But I believe that forcing the vars here should considerably narrow
the time window for such data races to occur.

Reviewed By: ndmitchell, grievejia

Differential Revision: D77883922

fbshipit-source-id: 5fb39166cfad5c68820727bc35182a247ba36393