Detect and panic with a dedicated error if we force `Var::ZERO`
Summary:
I've now run into this issue twice, it's pretty easy to forget about Var::ZERO
when forcing vars.

We should either produce a dedicated panic message (that's what this
diff does) or just skip the var. I opted for the dedicated panic because
it feels a little fishy to allow forcing vars that are just dummies, but
honestly I don't think it matters much - the important thing is to *avoid*
producing a misleading panic that says a `Var` escaped module boundaries,
because that can lead someone down the wrong debugging path.

Reviewed By: ndmitchell

Differential Revision: D77883921

fbshipit-source-id: 957373faa551a3908595217dd9405b0b2b7a5679