Add a test case tracking is_subset_eq issue
Summary:
We found a bug in enum logic but I think it's possible the same bug is going to
pop up elsewhere: if we see a metaclass that is a subtype of `Any`, an
`is_subset_eq` check against `EnumMetadata` will pass because that's how
gradual subtying works.

But we use `is_subset_eq` to decide "is the metaclass EnumMeta", and gradual
subtyping isn't really the right thing here - when handling metaclass
gymnastics in enum metadata checks, we aren't asking about assignability, we're
asking if we have plausible evidence that this is *actually* a subtype.

I think solving this is going to require either avoiding is_subset_eq entirely for
this kind of use case, or implementing an actual `is_subset_eq` that is distinct from
gradual assignability (which I know is a route `ty` has gone in); I'm not yet sure
the best approach.

Reviewed By: ndmitchell

Differential Revision: D77907516

fbshipit-source-id: 339475f665a938757d9b0b233e1877802bf35cf7