New test case for `type(...)` in base class list
Summary:
This is a missing feature; the gap in functionality is intended today, but
I suspect we need to eventually support it: some metaclasses are, in practice,
not easily available directly - for example a C extension might not expose it
in stubs.

There's an example of this in PyTorch code:
https://fburl.com/yk0j8i6u
Part of the example is not needed (`type(Generic)`, these days, is just
`type`!) but the bit using `type(torch._C.Future)` is likely actually
needed for runtime behavior.

Reviewed By: ndmitchell

Differential Revision: D77907515

fbshipit-source-id: ffd7a4add1202a975956b9383d3a3a7100cbe18d