Fix the EnumMeta handling with an Any base
Summary:
This partially addresses https://github.com/facebook/pyrefly/issues/622

This does not fully close the issue because there's still a question of whether
we should allow `type(X)` as the base class of a class (which is specifically useful
if you need a metaclass not available directly - something that can occur in C extensions).

But it does fix the clear-cut bug: we should
only treat a class as an enum if it actually *has* an
EnumMeta metaclass, not if its metaclass is assignable as an
`EnumMeta`. The `as_superclass` logic does what we need, it's
looking at the actual class hierarchy rather than using gradual
subtyping.

Reviewed By: ndmitchell

Differential Revision: D77907514

fbshipit-source-id: 9dfaf7c774f484c91fbf2b2a419ff456839174c5