Create interpreters.rs
Summary:
# This Diff

This diff moves interpreter fields from `ConfigFile` into a new dedicated, flattened struct. This is a pure refactor, and in upcoming diffs, we'll copy over functionality from other parts of `PythonEnvironment` related to finding/getting Python interpreters into this struct's implementation.

# Diff Stack

The upcoming diff stack fixes a bug with how we're finding and using interpreters. The way that it works now is actually slightly different from what the online documentation says, and now that I've had more time to think about it, I want it to behave slightly differently.

The next few diffs will fix the ordering and reorganize the logic around interpreter finding to be a bit cleaner.

Reviewed By: stroxler

Differential Revision: D77704773

fbshipit-source-id: 6f261f4cf86427e4aa4cea3b78d9e798807d6a2c