Move remaining find_interpreter logic into Interpreters
Summary:
# This Diff

This diff moves the remainder of the `find_interpreters()` logic from `ConfigFile::configure()` into `Interpreters::find_interpreters()`. This is mostly a refactor, but there is a signature change of the `Interpreters::find_interpreters()` function to return an `anyhow::Result<ConfigOrigin<PathBuf>>` instead of doing that conversion in `ConfigFile::configure()`.

# Diff Stack

The upcoming diff stack fixes a bug with how we're finding and using interpreters. The way that it works now is actually slightly different from what the online documentation says, and now that I've had more time to think about it, I want it to behave slightly differently.

The next few diffs will fix the ordering and reorganize the logic around interpreter finding to be a bit cleaner.

Reviewed By: grievejia

Differential Revision: D77752552

fbshipit-source-id: a090b74a5ecb4c384edbc900a720a6c30778c485