Make site_package_path() return an iterator like search_path()
Summary:
# Diff

This diff makes `ConfigFile::site_package_path()` return an iterator, similar to how `search_path()` does.

This will make it easy in the future to chain the upcoming `interpreter_site_package_path` to any user-provided `site_package_path`, without changing a lot of APIs to have to look at multiple search path lists.

# Context

This diff stack works on splitting the `site_package_path` into two fields, one for user-provided site packages and another for the site packages provided by the interpreter. We're doing this so that adding some values to the site package path won't completely overwrite it, and we can still insert an interpreter's site package path at the end of the list.

We want to do this because it's unlikely users would ever want to not use an interpreter's site package path if it exists. Once you begin adding specific values there, you must manually configure the entire thing, and we really want to avoid making users hard-code site package path if they can, since it make the config difficult to share between machines, especially if they're working on the same project.

Reviewed By: stroxler

Differential Revision: D77700215

fbshipit-source-id: 9731050bfc363fefcd3da7d5f30c492656e9054e