Always query interpreter unless explicitly instructed not to
Summary:
# Diff

This diff adds a new `--skip-interprerter-query` flag and config option, which are used in place of `PythonEnvironment::any_empty()` to determine when an interpreter should be queried.

We want to stop relying on `any_empty()` to figure out if we should query an interpreter, since we should be querying it and setting up the site package path *by default*. This way, explicitly adding stuff to `site_package_path` will enable configuring part of the site package path while still letting Pyrefly figure out the rest of it for you.

# Context

This diff stack works on splitting the `site_package_path` into two fields, one for user-provided site packages and another for the site packages provided by the interpreter. We're doing this so that adding some values to the site package path won't completely overwrite it, and we can still insert an interpreter's site package path at the end of the list.

We want to do this because it's unlikely users would ever want to not use an interpreter's site package path if it exists. Once you begin adding specific values there, you must manually configure the entire thing, and we really want to avoid making users hard-code site package path if they can, since it make the config difficult to share between machines, especially if they're working on the same project.

Reviewed By: stroxler

Differential Revision: D77699569

fbshipit-source-id: dacd10cef8386452c607c39cf8219713ec1bf437