Create ConfigOrigin enum to track where specific configs come from
Summary:
# This Diff

This diff adds a utility enum for tracking the source of a config value. I originally created it a while ago before deciding `SitePackagePathSource` was easier to do, but I have more fields that we need to keep track of the source of, so I'm actually gonna add it now.

This enum is a transparent value (which I think is kinda cool!). You can parse it into a value of its type without specifying it. Example:
```
// the struct
struct Test {
  a: Option<ConfigOption<String>>
}
// the toml file
[test]
a = "hello!"

// parsing it would come out as this:
Test { a: Some(ConfigOption::ConfigFile("hello!")) }
```

You can also serialize it without outputting the `ConfigOption` wrapper, and we're able to avoid serializing non-`ConfigFile` variants, allowing us to avoid dumping out values that are automatically configured that we might want to hide.

When it comes from anywhere else, we can explicitly set the value to something else (e.g. when we get a value from CLAP, we can set it on the config with another `ConfigOption` variant).

I also add in some helper functions around it similar to an `Option` type, which get used in upcoming diffs.

# Diff Stack

The upcoming diff stack fixes a bug with how we're finding and using interpreters. The way that it works now is actually slightly different from what the online documentation says, and now that I've had more time to think about it, I want it to behave slightly differently.

The next few diffs will fix the ordering and reorganize the logic around interpreter finding to be a bit cleaner.

Reviewed By: ndmitchell

Differential Revision: D77703777

fbshipit-source-id: 8771c54a0cde4de5930176944bb3a930ef58dd1c