Remove SitePackagePathSource
Summary:
# This Diff

With the addition of `ConfigOrigin`, `SitePackagePathSource` can be replaced by it. While looking at replacing it, I realized it doesn't really serve a purpose anymore with the `Interpreter` variant being removed, so I've decided to remove it and its uses completely.

`SitePackagePathSource`'s main purposes are to:
- let us know when a site package path entry is provided by the interpreter so that we don't do validation on values users didn't explicitly provide.
- output information on where site packages were found, which is really only useful when trying to figure out which interpreter it came from.

# Diff Stack

The upcoming diff stack fixes a bug with how we're finding and using interpreters. The way that it works now is actually slightly different from what the online documentation says, and now that I've had more time to think about it, I want it to behave slightly differently.

The next few diffs will fix the ordering and reorganize the logic around interpreter finding to be a bit cleaner.

Reviewed By: kinto0

Differential Revision: D77704771

fbshipit-source-id: 6797814d4dc5db8a34b63cead665a604beaa31df