Use ConfigOrigin to track where interpreter-specific config options come from
Summary:
# This Diff

This diff uses the recently created `ConfigOrigin` enum for `ConfigFile`'s `python_interpreter` and `conda_environment` config options.

When overwriting config values from CLI, we wrap those values in `ConfigOption::CommandLine` to appropriately mark where it came from, and we also use `ConfigOption::Auto` when the interpreter is determined with Pyrefly's config finding functionality.

# Diff Stack

The upcoming diff stack fixes a bug with how we're finding and using interpreters. The way that it works now is actually slightly different from what the online documentation says, and now that I've had more time to think about it, I want it to behave slightly differently.

The next few diffs will fix the ordering and reorganize the logic around interpreter finding to be a bit cleaner.

Reviewed By: grievejia

Differential Revision: D77703775

fbshipit-source-id: f9309946acfbda83f967c58904fd3edf923dc36c