Remove SitePackagePathSource::Interpreter
Summary:
# Diff

With a new field for `interpreter_site_package_path`, we have a better way to differentiate the search path entries. We can now remove the `SitePackagePathSource` variant for interpreter, since it's no longer useful.

# Context

This diff stack works on splitting the `site_package_path` into two fields, one for user-provided site packages and another for the site packages provided by the interpreter. We're doing this so that adding some values to the site package path won't completely overwrite it, and we can still insert an interpreter's site package path at the end of the list.

We want to do this because it's unlikely users would ever want to not use an interpreter's site package path if it exists. Once you begin adding specific values there, you must manually configure the entire thing, and we really want to avoid making users hard-code site package path if they can, since it make the config difficult to share between machines, especially if they're working on the same project.

Reviewed By: stroxler

Differential Revision: D77703773

fbshipit-source-id: ec66ff669f2930f30d5de81e2d1a5ec781d0ba9a