Support TypeVar in binop_infer
Summary:
TypeVar handling is missing from binop_infer. [issue https://github.com/facebook/pyrefly/issues/638]

I noticed that inside of `get_possible_attribute_bases`, we have `TypeVars` that are wrapped around a `Type`, and those aren't being visitited.

So in this diff, I recursivly visit the type, and create a quantified out of the TypeVar that's nested in `Type` and handle the `TypeVar`

RFC: can the attribute experts shed some light on whether this is the right approach of it's too much special casing?

Reviewed By: stroxler

Differential Revision: D77903461

fbshipit-source-id: 9599b3664ca4ce9b54aeaddd8b912ac9b887425f